Reading the main IPv4 address of a cluster host
The a cluster host have multiple IPs, "host" will return
several values.
If the main IP address of the cluster is IPv6,
the return value from this command will be the word "address".

Since this syntax for redis-trib does not support IPv6, this patch makes sure
we are getting the main IPv4 address of the host.

This fix #32